backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn==0.15.0
python-multipart>=0.0.18
pydantic>=2.9.0

# Database
sqlalchemy==1.4.41
alembic==1.7.1
psycopg2-binary==2.9.1

# AI/ML
spacy==3.7.4
transformers==4.38.0
tokenizers==0.15.2
torch>=2.2.0
numpy==1.22.0
pytesseract==0.3.8
python-docx==0.8.11
pdfminer.six==20201018

# Utils
passlib==1.7.4
python-dotenv==0.19.0
aiofiles==0.7.0
requests>=2.32.0
PyJWT==2.8.0

# Testing
pytest==6.2.5
pytest-asyncio==0.15.1
httpx>=0.23.0

# Logging and Monitoring
loguru==0.5.3
prometheus-client==0.11.0

# Email and Calendar
python-jose==3.3.0
google-auth>=2.28.1
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
python-dateutil>=2.8.2
email-validator>=2.1.0.post1
aiosmtplib>=3.0.1

# Before installing spacy models, run:
# python -m spacy download en_core_web_lg